requirements.txt
fastapi==0.111.0
python-dotenv==1.0.1
uvicorn==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4
arq==0.26.0
redis<5.0.0,>=4.2.0
jsonschema==4.23.0
httpx==0.27.0
loguru==0.7.2
google-genai>=0.8.0
qdrant-client>=1.9.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
openai>=1.0.0
neo4j>=5.21.0
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.3
